+++ sys/arch/x86/include/pci_machdep.h
@@ -36,7 +36,7 @@
 /*
  * Types provided to machine-independent PCI code
  */
-typedef int pci_intr_handle_t;
+typedef uint64_t pci_intr_handle_t;
 
 #include <x86/pci_machdep_common.h>
